backend/requirements.txt
# Smart Interview Simulator - Backend Requirements
# Python dependencies for the FastAPI backend

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# CORS
fastapi-cors==0.0.1

# Database
pymongo==4.6.1

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# NLP Processing
spacy==3.7.4
nltk==3.8.1

# Machine Learning (optional, for advanced NLP)
scikit-learn==1.4.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Install spaCy model: python -m spacy download en_core_web_sm